Renamed to prevent conflicts with other libraries
