+++ app/notification.py
@@ -3,10 +3,10 @@ Handles sending notifications via the configured notifiers
 """
 
 import structlog
-from notifiers.twilio import TwilioNotifier
-from notifiers.slack import SlackNotifier
-from notifiers.gmail import GmailNotifier
-from notifiers.integram import IntegramNotifier
+from notifiers.twilio_client import TwilioNotifier
+from notifiers.slack_client import SlackNotifier
+from notifiers.gmail_client import GmailNotifier
+from notifiers.integram_client import IntegramNotifier
 
 class Notifier():
     """
